Refactor code for conditional compilation across UEFI and Host Targets
This refactor ensures that the paging crate is buildable for both UEFI and host targets.
It introduces appropriate conditional compilation to support:

- **UEFI Targets**: `x86_64-unknown-uefi`, `aarch64-unknown-uefi`
- **Host Targets**: `x86_64-pc-windows-msvc`, `aarch64-pc-windows-msvc`, `x86_64-unknown-linux-gnu`, `aarch64-unknown-linux-gnu`

More details can be found in `src\CODE.md`.
