+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "paging"
-version = "4.0.0"
+version = "4.0.1"
 publish = ["UefiRust"]
 edition = "2021"
 license = "BSD-2-Clause-Patent"
@@ -15,7 +15,7 @@ features = ["doc"]
 bitfield-struct = "0.10"
 bitflags = "2.6.0"
 log = { version = "^0.4", default-features = false }
+cfg-if = "1.0.0"
 
 [features]
-std = []
 doc = []
