+++ src/aarch64/tests.rs
@@ -1,4 +1,2 @@
 mod aarch64_paging_tests;
 mod aarch64_test_page_allocator;
-mod x64_paging_tests;
-mod x64_test_page_allocator;
