+++ src/x64/paging.rs
@@ -867,160 +867,3 @@ impl<A: PageAllocator> PageTable for X64PageTable<A> {
         log::info!("{}", "-".repeat(132));
     }
 }
-
-/// Given the [start, end offset] at the current level from the [start, end VA],
-/// this function calculates the number of entries required for the range. It
-/// considers the number of entries at the parent level because the start and
-/// end offsets might span across multiple pages.
-fn find_num_entries(start_offset: u64, end_offset: u64, num_entries_at_parent_level: u64) -> u64 {
-    let mut num_entries = 0;
-
-    // Entries spanning multiple pages
-    if num_entries_at_parent_level > 1 {
-        num_entries += 512 - start_offset; // Number of upper entries in first page
-        num_entries += (num_entries_at_parent_level - 2) * 512; // number of entries in between pages
-        num_entries += end_offset + 1; // Number of lower entries in the last page
-    } else {
-        // Entries do not span multiple pages(end_offset is guaranteed to be higher than start offset)
-        num_entries = end_offset - start_offset + 1; // Number of entries in the page
-    }
-
-    num_entries
-}
-
-/// Finds the number of pages that are saved using large pages for the given address
-/// range and paging levels compared for using the lowest level.
-fn find_large_page_savings(address: u64, size: u64, level: PageLevel, lowest_page_level: PageLevel) -> u64 {
-    // The number of large pages in a given address start & length is deterministic
-    // based on the alignment of the address to the individual large pages size.
-    // Recurse down through levels finding the optimal page size to use.
-
-    if (level == lowest_page_level) || (size == 0) {
-        return 0;
-    }
-
-    if !level.supports_pa_entry() {
-        return find_large_page_savings(address, size, level - 1, lowest_page_level);
-    }
-
-    let mut savings = 0;
-    let alignment = level.entry_va_size();
-    let aligned_address = (address + alignment - 1) & !(alignment - 1);
-
-    // If there are no large pages that can be used for the given address range,
-    // then continue with the next level.
-    if aligned_address + alignment > address + size {
-        return find_large_page_savings(address, size, level - 1, lowest_page_level);
-    }
-
-    // Split of the unaligned beginning and end and recursive to the next level
-    // to find the savings with smaller page sizes.
-
-    savings += find_large_page_savings(address, aligned_address - address, level - 1, lowest_page_level);
-
-    let aligned_end = (address + size) & !(alignment - 1);
-    let remainder = (address + size) - aligned_end;
-
-    savings += find_large_page_savings(aligned_end, remainder, level - 1, lowest_page_level);
-
-    // The savings is the number of sub-pages that would be saved by each large page
-    // which is 1 for the current level and then 512 for each level below which.
-    // e.g. a large page at the third level would save 1 + 512
-    let num_large_pages = (aligned_end - aligned_address) / alignment;
-    let page_entries: u64 = 512;
-    let remaining_levels = level as u64 - lowest_page_level as u64;
-
-    savings += num_large_pages;
-    if remaining_levels > 1 {
-        savings += num_large_pages * page_entries.pow(remaining_levels as u32 - 1);
-    }
-
-    savings
-}
-
-pub(crate) fn num_page_tables_required(address: u64, size: u64, paging_type: PagingType) -> PtResult<u64> {
-    let address = VirtualAddress::new(address);
-    if size == 0 || !address.is_4kb_aligned() {
-        return Err(PtError::UnalignedAddress);
-    }
-
-    // Check the memory range is aligned
-    if !(address + size).is_4kb_aligned() {
-        return Err(PtError::UnalignedAddress);
-    }
-
-    let start_va = address;
-    let end_va = address + size - 1;
-
-    // For the given paging type, identify the highest and lowest page levels.
-    // This is used during page building to terminate the recursion.
-    let (highest_page_level, lowest_page_level) = match paging_type {
-        PagingType::Paging5Level => (PageLevel::Pml5, PageLevel::Pt),
-        PagingType::Paging4Level => (PageLevel::Pml4, PageLevel::Pt),
-        _ => return Err(PtError::InvalidParameter),
-    };
-
-    let mut num_entries_at_parent_level = 0;
-    let mut num_tables_at_current_level = 1; // top level table
-    let mut total_num_tables = 0;
-
-    // Rust does not support creating ranges [high..=low], so we use
-    // [low..=high].rev() instead.
-    for level in ((lowest_page_level as u64)..=(highest_page_level as u64)).rev() {
-        // Add the number of tables required at the current level to the total
-        // pages. This has already been computed in the previous iteration.
-        total_num_tables += num_tables_at_current_level;
-
-        let start_offset = start_va.get_index(level.into());
-        let end_offset = end_va.get_index(level.into());
-
-        // Prepare for the next level: Calculating the number of tables required
-        // at the next level (e.g., PDP) depends on the number of entries
-        // present at the current level (e.g., PML4). Calculating the number of
-        // entries at the current level (PML4) in turn depends on the number of
-        // entries at the parent level (PML5 — this is the third parameter).
-        // Why? See below.
-
-        //  |  parent level |  current level |  next level
-        //  |               |                |
-        //  │               │  ┌─────┐       │
-        //  │               │  │     │       │
-        //  │               │  ├─────┤       │
-        //  │               │  │     │       │
-        //  │               │  ├─────┤       │
-        //  │               └─►│PML4E│       │
-        //  │               │  ├─────┤       │
-        //  │               │  │PML4E|       │
-        //  │          ┌──────►└─────┘       │
-        //  │          │    │  ┌─────┐       │  ┌─────┐
-        //  │          │    │  │PML4E│       │  │     │
-        //  │          │    │  ├─────┤       │  ├─────┤
-        //  │          │    │  │PML4E│       │  │     │
-        //  │          │    │  ├─────┤       │  ├─────┤
-        //  │          │    └─►│PML4E│       │  │PDPE │
-        //  │          │    │  ├─────┤       │  ├─────┤
-        //  │          │    │  │PML4E|       |  │PDPE |
-        //  │          │ ┌────►└─────┘   ┌─────►└─────┘
-        //  │  ┌─────┐ │ │  │  ┌─────┐   │   │  ┌─────┐
-        //  │  │PML5E│─┘ │  │  │PML4E|───┘   │  │PDPE |
-        //  │  ├─────┤   │  │  ├─────┤       │  ├─────┤
-        //  │  │PML5E│───┘  └─►│PML4E│───┐   │  │PDPE │
-        //  │  ├─────┤         ├─────┤   │   │  ├─────┤
-        //  └─►│PML5E├───┐     │     │   │   └─►│PDPE │───┐
-        //     ├─────┤   │     ├─────┤   │      ├─────┤   │
-        //     │     │   │     │     │   │      │     │   │
-        //     └─────┘   └────►└─────┘   └─────►└─────┘   └───►
-        let num_entries_at_current_level = find_num_entries(start_offset, end_offset, num_entries_at_parent_level);
-
-        // These are truly consumed in the next iteration.
-        num_tables_at_current_level = num_entries_at_current_level;
-        num_entries_at_parent_level = num_entries_at_current_level;
-    }
-
-    // The above calculates only the lowest pages, now calculate saving through large
-    // pages.
-    let savings = find_large_page_savings(address.into(), size, highest_page_level, lowest_page_level);
-    total_num_tables -= savings;
-
-    Ok(total_num_tables)
-}
