+++ src/x64/tests.rs
@@ -1,0 +1,2 @@
+mod x64_paging_tests;
+mod x64_test_page_allocator;
