+++ src/x64/tests/x64_paging_tests.rs
@@ -1,74 +1,257 @@
+use log::{Level, LevelFilter, Metadata, Record};
+
+use crate::PageTable;
 use crate::{
-    aarch64::{
-        paging::{num_page_tables_required, AArch64PageTable},
-        structs::*,
+    x64::{
+        structs::{PageTableEntry, CR3_PAGE_BASE_ADDRESS_MASK, FRAME_SIZE_4KB, SELF_MAP_INDEX, ZERO_VA_INDEX},
+        tests::x64_test_page_allocator::TestPageAllocator,
+        X64PageTable,
     },
-    tests::aarch64_test_page_allocator::TestPageAllocator,
-    MemoryAttributes, PageTable, PagingType, PtError,
+    MemoryAttributes, PagingType, PtError,
 };
+
+// Sample logger for log crate to dump stuff in tests
+struct SimpleLogger;
+
+impl log::Log for SimpleLogger {
+    fn enabled(&self, metadata: &Metadata) -> bool {
+        metadata.level() <= Level::Info
+    }
+
+    fn log(&self, record: &Record) {
+        if self.enabled(record.metadata()) {
+            println!("{}", record.args());
+        }
+    }
+
+    fn flush(&self) {}
+}
+static LOGGER: SimpleLogger = SimpleLogger;
+
+fn set_logger() {
+    let _ = log::set_logger(&LOGGER).map(|()| log::set_max_level(LevelFilter::Info));
+}
+
+/// Given the [start, end offset] at the current level from the [start, end VA],
+/// this function calculates the number of entries required for the range. It
+/// considers the number of entries at the parent level because the start and
+/// end offsets might span across multiple pages.
+fn find_num_entries(start_offset: u64, end_offset: u64, num_entries_at_parent_level: u64) -> u64 {
+    let mut num_entries = 0;
+
+    // Entries spanning multiple pages
+    if num_entries_at_parent_level > 1 {
+        num_entries += 512 - start_offset; // Number of upper entries in first page
+        num_entries += (num_entries_at_parent_level - 2) * 512; // number of entries in between pages
+        num_entries += end_offset + 1; // Number of lower entries in the last page
+    } else {
+        // Entries do not span multiple pages(end_offset is guaranteed to be higher than start offset)
+        num_entries = end_offset - start_offset + 1; // Number of entries in the page
+    }
+
+    num_entries
+}
+
+/// Finds the number of pages that are saved using large pages for the given address
+/// range and paging levels compared for using the lowest level.
+fn find_large_page_savings(address: u64, size: u64, level: PageLevel, lowest_page_level: PageLevel) -> u64 {
+    // The number of large pages in a given address start & length is deterministic
+    // based on the alignment of the address to the individual large pages size.
+    // Recurse down through levels finding the optimal page size to use.
+
+    if (level == lowest_page_level) || (size == 0) {
+        return 0;
+    }
+
+    if !level.supports_pa_entry() {
+        return find_large_page_savings(address, size, level - 1, lowest_page_level);
+    }
+
+    let mut savings = 0;
+    let alignment = level.entry_va_size();
+    let aligned_address = (address + alignment - 1) & !(alignment - 1);
+
+    // If there are no large pages that can be used for the given address range,
+    // then continue with the next level.
+    if aligned_address + alignment > address + size {
+        return find_large_page_savings(address, size, level - 1, lowest_page_level);
+    }
+
+    // Split of the unaligned beginning and end and recursive to the next level
+    // to find the savings with smaller page sizes.
+
+    savings += find_large_page_savings(address, aligned_address - address, level - 1, lowest_page_level);
+
+    let aligned_end = (address + size) & !(alignment - 1);
+    let remainder = (address + size) - aligned_end;
+
+    savings += find_large_page_savings(aligned_end, remainder, level - 1, lowest_page_level);
+
+    // The savings is the number of sub-pages that would be saved by each large page
+    // which is 1 for the current level and then 512 for each level below which.
+    // e.g. a large page at the third level would save 1 + 512
+    let num_large_pages = (aligned_end - aligned_address) / alignment;
+    let page_entries: u64 = 512;
+    let remaining_levels = level as u64 - lowest_page_level as u64;
+
+    savings += num_large_pages;
+    if remaining_levels > 1 {
+        savings += num_large_pages * page_entries.pow(remaining_levels as u32 - 1);
+    }
+
+    savings
+}
+
+fn num_page_tables_required(address: u64, size: u64, paging_type: PagingType) -> PtResult<u64> {
+    let address = VirtualAddress::new(address);
+    if size == 0 || !address.is_4kb_aligned() {
+        return Err(PtError::UnalignedAddress);
+    }
+
+    // Check the memory range is aligned
+    if !(address + size).is_4kb_aligned() {
+        return Err(PtError::UnalignedAddress);
+    }
+
+    let start_va = address;
+    let end_va = address + size - 1;
+
+    // For the given paging type, identify the highest and lowest page levels.
+    // This is used during page building to terminate the recursion.
+    let (highest_page_level, lowest_page_level) = match paging_type {
+        PagingType::Paging5Level => (PageLevel::Pml5, PageLevel::Pt),
+        PagingType::Paging4Level => (PageLevel::Pml4, PageLevel::Pt),
+        _ => return Err(PtError::InvalidParameter),
+    };
+
+    let mut num_entries_at_parent_level = 0;
+    let mut num_tables_at_current_level = 1; // top level table
+    let mut total_num_tables = 0;
+
+    // Rust does not support creating ranges [high..=low], so we use
+    // [low..=high].rev() instead.
+    for level in ((lowest_page_level as u64)..=(highest_page_level as u64)).rev() {
+        // Add the number of tables required at the current level to the total
+        // pages. This has already been computed in the previous iteration.
+        total_num_tables += num_tables_at_current_level;
+
+        let start_offset = start_va.get_index(level.into());
+        let end_offset = end_va.get_index(level.into());
+
+        // Prepare for the next level: Calculating the number of tables required
+        // at the next level (e.g., PDP) depends on the number of entries
+        // present at the current level (e.g., PML4). Calculating the number of
+        // entries at the current level (PML4) in turn depends on the number of
+        // entries at the parent level (PML5 — this is the third parameter).
+        // Why? See below.
+
+        //  |  parent level |  current level |  next level
+        //  |               |                |
+        //  │               │  ┌─────┐       │
+        //  │               │  │     │       │
+        //  │               │  ├─────┤       │
+        //  │               │  │     │       │
+        //  │               │  ├─────┤       │
+        //  │               └─►│PML4E│       │
+        //  │               │  ├─────┤       │
+        //  │               │  │PML4E|       │
+        //  │          ┌──────►└─────┘       │
+        //  │          │    │  ┌─────┐       │  ┌─────┐
+        //  │          │    │  │PML4E│       │  │     │
+        //  │          │    │  ├─────┤       │  ├─────┤
+        //  │          │    │  │PML4E│       │  │     │
+        //  │          │    │  ├─────┤       │  ├─────┤
+        //  │          │    └─►│PML4E│       │  │PDPE │
+        //  │          │    │  ├─────┤       │  ├─────┤
+        //  │          │    │  │PML4E|       |  │PDPE |
+        //  │          │ ┌────►└─────┘   ┌─────►└─────┘
+        //  │  ┌─────┐ │ │  │  ┌─────┐   │   │  ┌─────┐
+        //  │  │PML5E│─┘ │  │  │PML4E|───┘   │  │PDPE |
+        //  │  ├─────┤   │  │  ├─────┤       │  ├─────┤
+        //  │  │PML5E│───┘  └─►│PML4E│───┐   │  │PDPE │
+        //  │  ├─────┤         ├─────┤   │   │  ├─────┤
+        //  └─►│PML5E├───┐     │     │   │   └─►│PDPE │───┐
+        //     ├─────┤   │     ├─────┤   │      ├─────┤   │
+        //     │     │   │     │     │   │      │     │   │
+        //     └─────┘   └────►└─────┘   └─────►└─────┘   └───►
+        let num_entries_at_current_level = find_num_entries(start_offset, end_offset, num_entries_at_parent_level);
+
+        // These are truly consumed in the next iteration.
+        num_tables_at_current_level = num_entries_at_current_level;
+        num_entries_at_parent_level = num_entries_at_current_level;
+    }
+
+    // The above calculates only the lowest pages, now calculate saving through large
+    // pages.
+    let savings = find_large_page_savings(address.into(), size, highest_page_level, lowest_page_level);
+    total_num_tables -= savings;
+
+    Ok(total_num_tables)
+}
 
 #[test]
 fn test_find_num_page_tables() {
-    // Mapping one page of physical address require 4 page tables.
+    // Mapping one page of physical address require 4 page tables(PML4/PDP/PD/PT)
     let address = 0x0;
     let size = FRAME_SIZE_4KB; // 4k
-    let res = num_page_tables_required(address, size, PagingType::AArch64PageTable4KB);
+    let res = num_page_tables_required(address, size, PagingType::Paging4Level);
     assert!(res.is_ok());
     let table_count = res.unwrap();
     assert_eq!(table_count, 4);
 
-    // Mapping 511 pages of physical address require 4 page tables.
+    // Mapping 511 pages of physical address require 4 page tables(PML4/PDP/PD/PT)
     let address = FRAME_SIZE_4KB;
     let size = 511 * FRAME_SIZE_4KB;
-    let res = num_page_tables_required(address, size, PagingType::AArch64PageTable4KB);
+    let res = num_page_tables_required(address, size, PagingType::Paging4Level);
     assert!(res.is_ok());
     let table_count = res.unwrap();
     assert_eq!(table_count, 4);
 
-    // Mapping 512 pages of physical address require 3 page tables due to 2MB large pages.
+    // Mapping 512 pages of physical address require 3 page tables because of 2mb pages.(PML4/PDP/PD)
     let address = 0x0;
     let size = 512 * FRAME_SIZE_4KB;
-    let res = num_page_tables_required(address, size, PagingType::AArch64PageTable4KB);
+    let res = num_page_tables_required(address, size, PagingType::Paging4Level);
     assert!(res.is_ok());
     let table_count = res.unwrap();
     assert_eq!(table_count, 3);
 
-    // Mapping 513 pages of physical address require 4 page tables due to 1 2MB page and 1 4KB page.
+    // Mapping 513 pages of physical address require 4 page tables because it will be 1 2mb mapping and 1 4kb.
+    // (PML5(1)/PML4(1)/PDPE(1)/PDP(1)/PT(1))
     let address = 0x0;
-    let size = 513 * FRAME_SIZE_4KB + FRAME_SIZE_4KB;
-    let res = num_page_tables_required(address, size, PagingType::AArch64PageTable4KB);
+    let size = 513 * FRAME_SIZE_4KB;
+    let res = num_page_tables_required(address, size, PagingType::Paging4Level);
     assert!(res.is_ok());
     let table_count = res.unwrap();
     assert_eq!(table_count, 4);
 
-    // Mapping 1Gb pages of physical address require 2 page tables due to 1Gb large pages.
+    // Mapping 1gb of physical address require 2 page tables because of 1Gb pages.(PML4/PDP)
     let address = 0x0;
     let size = 512 * 512 * FRAME_SIZE_4KB;
-    let res = num_page_tables_required(address, size, PagingType::AArch64PageTable4KB);
+    let res = num_page_tables_required(address, size, PagingType::Paging4Level);
     assert!(res.is_ok());
     let table_count = res.unwrap();
     assert_eq!(table_count, 2);
 
-    // Mapping 1 1Gb Page + 1 2mb page + 1 4kb page require 4 page tables.
+    // Mapping 1 1GbPage + 1 2mb page + 1 4kb page require 4 page tables.(PML4/PDP/PD/PT)
     let address = 0x0;
     let size = (512 * 512 * FRAME_SIZE_4KB) + (512 * FRAME_SIZE_4KB) + FRAME_SIZE_4KB;
-    let res = num_page_tables_required(address, size, PagingType::AArch64PageTable4KB);
+    let res = num_page_tables_required(address, size, PagingType::Paging4Level);
     assert!(res.is_ok());
     let table_count = res.unwrap();
     assert_eq!(table_count, 4);
 
-    // Mapping 2mb starting at 2mb/2 should take 5 pages.
+    // Mapping 2mb starting at 2mb/2 should take 5 pages. (PML4/PDP/PD(1)/PT(2))
     let address = 256 * FRAME_SIZE_4KB;
     let size = 512 * FRAME_SIZE_4KB;
-    let res = num_page_tables_required(address, size, PagingType::AArch64PageTable4KB);
+    let res = num_page_tables_required(address, size, PagingType::Paging4Level);
     assert!(res.is_ok());
     let table_count = res.unwrap();
     assert_eq!(table_count, 5);
 
-    // Mapping 10Gb starting at 4kb should take 6 pages.
+    // Mapping 10Gb starting at 4kb should take 6 pages. (PML4/PDP/PD(2)/PT(2))
     let address = FRAME_SIZE_4KB;
     let size = 10 * 512 * 512 * FRAME_SIZE_4KB;
-    let res = num_page_tables_required(address, size, PagingType::AArch64PageTable4KB);
+    let res = num_page_tables_required(address, size, PagingType::Paging4Level);
     assert!(res.is_ok());
     let table_count = res.unwrap();
     assert_eq!(table_count, 6);
@@ -84,54 +267,25 @@ fn test_map_memory_address_simple() {
         size: u64,
     }
 
-    let test_configs = [TestConfig { paging_type: PagingType::AArch64PageTable4KB, address: 0, size: 0x400000 }];
+    let test_configs = [
+        TestConfig { paging_type: PagingType::Paging4Level, address: 0, size: 0x400000 },
+        TestConfig { paging_type: PagingType::Paging5Level, address: 0, size: 0x400000 },
+    ];
 
     for test_config in test_configs {
         let TestConfig { size, address, paging_type } = test_config;
 
-        let num_pages = num_page_tables_required(address, size, paging_type).unwrap() + 0x3;
+        let num_pages = num_page_tables_required(address, size, paging_type).unwrap() + 0x4;
 
         let page_allocator = TestPageAllocator::new(num_pages, paging_type);
-
-        let pt = AArch64PageTable::new(page_allocator.clone(), paging_type);
+        let pt = X64PageTable::new(page_allocator.clone(), paging_type);
 
         assert!(pt.is_ok());
         let mut pt = pt.unwrap();
 
-        let attributes = MemoryAttributes::ReadOnly | MemoryAttributes::WriteCombining;
+        let attributes = MemoryAttributes::ReadOnly;
         let res = pt.map_memory_region(address, size, attributes);
-        assert!(res.is_ok());
 
-        assert_eq!(page_allocator.pages_allocated(), num_pages);
-
-        page_allocator.validate_pages(address, size, attributes);
-    }
-}
-
-#[test]
-fn test_map_memory_address_not_so_simple() {
-    struct TestConfig {
-        paging_type: PagingType,
-        address: u64,
-        size: u64,
-    }
-
-    let test_configs = [TestConfig { paging_type: PagingType::AArch64PageTable4KB, address: 0, size: 0x400000 }];
-
-    for test_config in test_configs {
-        let TestConfig { size, address, paging_type } = test_config;
-
-        let num_pages = num_page_tables_required(address, size, paging_type).unwrap() + 0x3;
-
-        let page_allocator = TestPageAllocator::new(num_pages, paging_type);
-
-        let pt = AArch64PageTable::new(page_allocator.clone(), paging_type);
-
-        assert!(pt.is_ok());
-        let mut pt = pt.unwrap();
-
-        let attributes = MemoryAttributes::ReadOnly | MemoryAttributes::Uncacheable;
-        let res = pt.map_memory_region(address, size, attributes);
         assert!(res.is_ok());
 
         assert_eq!(page_allocator.pages_allocated(), num_pages);
@@ -148,28 +302,28 @@ fn test_map_memory_address_0_to_ffff_ffff() {
         size: u64,
     }
 
-    let test_configs = [TestConfig { paging_type: PagingType::AArch64PageTable4KB, address: 0, size: FRAME_SIZE_4KB }];
+    let test_configs = [
+        TestConfig { paging_type: PagingType::Paging4Level, address: 0, size: FRAME_SIZE_4KB },
+        TestConfig { paging_type: PagingType::Paging5Level, address: 0, size: FRAME_SIZE_4KB },
+    ];
 
     for test_config in test_configs {
         let TestConfig { mut size, address, paging_type } = test_config;
 
         while size < 0xffff_ffff {
-            let num_pages = num_page_tables_required(address, size, paging_type).unwrap() + 0x3;
+            let num_pages = num_page_tables_required(address, size, paging_type).unwrap() + 0x4;
 
             let page_allocator = TestPageAllocator::new(num_pages, paging_type);
-
-            let pt = AArch64PageTable::new(page_allocator.clone(), paging_type);
+            let pt = X64PageTable::new(page_allocator.clone(), paging_type);
 
             assert!(pt.is_ok());
             let mut pt = pt.unwrap();
 
-            let attributes = MemoryAttributes::ReadOnly | MemoryAttributes::WriteCombining;
+            let attributes = MemoryAttributes::ReadOnly;
             let res = pt.map_memory_region(address, size, attributes);
-            if res.is_err() {
-                log::info!("addressW: {:x} size: {:x}", address, size);
-            }
             assert!(res.is_ok());
 
+            log::info!("allocated: {} expected: {}", page_allocator.pages_allocated(), num_pages);
             assert_eq!(page_allocator.pages_allocated(), num_pages);
 
             page_allocator.validate_pages(address, size, attributes);
@@ -189,26 +343,33 @@ fn test_map_memory_address_single_page_from_0_to_ffff_ffff() {
         address_increment: u64,
     }
 
-    let test_configs = [TestConfig {
-        paging_type: PagingType::AArch64PageTable4KB,
-        address: 0,
-        address_increment: FRAME_SIZE_4KB,
-        size: FRAME_SIZE_4KB,
-    }];
+    let test_configs = [
+        TestConfig {
+            paging_type: PagingType::Paging4Level,
+            address: 0,
+            address_increment: FRAME_SIZE_4KB,
+            size: FRAME_SIZE_4KB,
+        },
+        TestConfig {
+            paging_type: PagingType::Paging5Level,
+            address: 0,
+            address_increment: FRAME_SIZE_4KB,
+            size: FRAME_SIZE_4KB,
+        },
+    ];
 
     for test_config in test_configs {
         let TestConfig { size, mut address, address_increment, paging_type } = test_config;
         while address < 0xffff_ffff {
-            let num_pages = num_page_tables_required(address, size, paging_type).unwrap() + 0x3;
+            let num_pages = num_page_tables_required(address, size, paging_type).unwrap() + 0x4;
 
             let page_allocator = TestPageAllocator::new(num_pages, paging_type);
-
-            let pt = AArch64PageTable::new(page_allocator.clone(), paging_type);
+            let pt = X64PageTable::new(page_allocator.clone(), paging_type);
 
             assert!(pt.is_ok());
             let mut pt = pt.unwrap();
 
-            let attributes = MemoryAttributes::ReadOnly | MemoryAttributes::WriteCombining;
+            let attributes = MemoryAttributes::ReadOnly;
             let res = pt.map_memory_region(address, size, attributes);
             assert!(res.is_ok());
 
@@ -229,27 +390,34 @@ fn test_map_memory_address_multiple_page_from_0_to_ffff_ffff() {
         address_increment: u64,
     }
 
-    let test_configs = [TestConfig {
-        paging_type: PagingType::AArch64PageTable4KB,
-        address: 0,
-        address_increment: FRAME_SIZE_4KB,
-        size: FRAME_SIZE_4KB << 1,
-    }];
+    let test_configs = [
+        TestConfig {
+            paging_type: PagingType::Paging4Level,
+            address: 0,
+            address_increment: FRAME_SIZE_4KB,
+            size: FRAME_SIZE_4KB << 1,
+        },
+        TestConfig {
+            paging_type: PagingType::Paging5Level,
+            address: 0,
+            address_increment: FRAME_SIZE_4KB,
+            size: FRAME_SIZE_4KB << 1,
+        },
+    ];
 
     for test_config in test_configs {
         let TestConfig { size, mut address, address_increment, paging_type } = test_config;
 
         while address < 0xffff_ffff {
-            let num_pages = num_page_tables_required(address, size, paging_type).unwrap() + 0x3;
+            let num_pages = num_page_tables_required(address, size, paging_type).unwrap() + 0x4;
 
             let page_allocator = TestPageAllocator::new(num_pages, paging_type);
-
-            let pt = AArch64PageTable::new(page_allocator.clone(), paging_type);
+            let pt = X64PageTable::new(page_allocator.clone(), paging_type);
 
             assert!(pt.is_ok());
             let mut pt = pt.unwrap();
 
-            let attributes = MemoryAttributes::ReadOnly | MemoryAttributes::WriteCombining;
+            let attributes = MemoryAttributes::ReadOnly;
             let res = pt.map_memory_region(address, size, attributes);
             assert!(res.is_ok());
 
@@ -269,7 +437,10 @@ fn test_map_memory_address_unaligned() {
         size: u64,
     }
 
-    let test_configs = [TestConfig { paging_type: PagingType::AArch64PageTable4KB, address: 0x1, size: 200 }];
+    let test_configs = [
+        TestConfig { paging_type: PagingType::Paging4Level, address: 0x1, size: 200 },
+        TestConfig { paging_type: PagingType::Paging5Level, address: 0x1, size: 200 },
+    ];
 
     for test_config in test_configs {
         let TestConfig { size, address, paging_type } = test_config;
@@ -277,13 +448,12 @@ fn test_map_memory_address_unaligned() {
         let max_pages: u64 = 10;
 
         let page_allocator = TestPageAllocator::new(max_pages, paging_type);
-
-        let pt = AArch64PageTable::new(page_allocator.clone(), paging_type);
+        let pt = X64PageTable::new(page_allocator.clone(), paging_type);
 
         assert!(pt.is_ok());
         let mut pt = pt.unwrap();
 
-        let attributes = MemoryAttributes::ReadOnly | MemoryAttributes::WriteCombining;
+        let attributes = MemoryAttributes::ReadOnly;
         let res = pt.map_memory_region(address, size, attributes);
         assert!(res.is_err());
         assert_eq!(res, Err(PtError::UnalignedAddress));
@@ -300,7 +470,8 @@ fn test_map_memory_address_range_overflow() {
 
     let test_configs = [
         // VA range overflows
-        TestConfig { paging_type: PagingType::AArch64PageTable4KB, address: MAX_VA, size: MAX_VA },
+        TestConfig { paging_type: PagingType::Paging4Level, address: 0xffff_ffff_ffff_f000, size: 0x2000 },
+        TestConfig { paging_type: PagingType::Paging5Level, address: 0xffff_ffff_ffff_f000, size: 0x2000 },
     ];
 
     for test_config in test_configs {
@@ -309,43 +480,12 @@ fn test_map_memory_address_range_overflow() {
         let max_pages: u64 = 10;
 
         let page_allocator = TestPageAllocator::new(max_pages, paging_type);
-        let pt = AArch64PageTable::new(page_allocator.clone(), paging_type);
+        let pt = X64PageTable::new(page_allocator.clone(), paging_type);
 
         assert!(pt.is_ok());
         let mut pt = pt.unwrap();
 
-        let attributes = MemoryAttributes::ReadOnly | MemoryAttributes::WriteCombining;
-        let res = pt.map_memory_region(address, size, attributes);
-        assert!(res.is_err());
-        assert_eq!(res, Err(PtError::InvalidMemoryRange));
-    }
-}
-
-#[test]
-fn test_map_memory_address_invalid_range() {
-    struct TestConfig {
-        paging_type: PagingType,
-        address: u64,
-        size: u64,
-    }
-
-    let test_configs = [
-        // VA above the valid address range
-        TestConfig { paging_type: PagingType::AArch64PageTable4KB, address: MAX_VA + 1, size: FRAME_SIZE_4KB },
-    ];
-
-    for test_config in test_configs {
-        let TestConfig { size, address, paging_type } = test_config;
-
-        let max_pages: u64 = 10;
-
-        let page_allocator = TestPageAllocator::new(max_pages, paging_type);
-        let pt = AArch64PageTable::new(page_allocator.clone(), paging_type);
-
-        assert!(pt.is_ok());
-        let mut pt = pt.unwrap();
-
-        let attributes = MemoryAttributes::ReadOnly | MemoryAttributes::WriteCombining;
+        let attributes = MemoryAttributes::ReadOnly;
         let res = pt.map_memory_region(address, size, attributes);
         assert!(res.is_err());
         assert_eq!(res, Err(PtError::InvalidMemoryRange));
@@ -360,7 +500,10 @@ fn test_map_memory_address_zero_size() {
         size: u64,
     }
 
-    let test_configs = [TestConfig { paging_type: PagingType::AArch64PageTable4KB, address: 0x1000, size: 0 }];
+    let test_configs = [
+        TestConfig { paging_type: PagingType::Paging4Level, address: 0x1000, size: 0 },
+        TestConfig { paging_type: PagingType::Paging5Level, address: 0x1000, size: 0 },
+    ];
 
     for test_config in test_configs {
         let TestConfig { size, address, paging_type } = test_config;
@@ -368,13 +511,12 @@ fn test_map_memory_address_zero_size() {
         let max_pages: u64 = 10;
 
         let page_allocator = TestPageAllocator::new(max_pages, paging_type);
-
-        let pt = AArch64PageTable::new(page_allocator.clone(), paging_type);
+        let pt = X64PageTable::new(page_allocator.clone(), paging_type);
 
         assert!(pt.is_ok());
         let mut pt = pt.unwrap();
 
-        let attributes = MemoryAttributes::ReadOnly | MemoryAttributes::WriteCombining;
+        let attributes = MemoryAttributes::ReadOnly;
         let res = pt.map_memory_region(address, size, attributes);
         assert!(res.is_err());
         assert_eq!(res, Err(PtError::InvalidMemoryRange));
@@ -391,25 +533,23 @@ fn test_unmap_memory_address_simple() {
         size: u64,
     }
 
-    let test_configs = [TestConfig {
-        paging_type: PagingType::AArch64PageTable4KB,
-        address: 0x1000,
-        size: FRAME_SIZE_4KB * 512 * 512 * 10,
-    }];
+    let test_configs = [
+        TestConfig { paging_type: PagingType::Paging4Level, address: 0x1000, size: FRAME_SIZE_4KB * 512 * 512 * 10 },
+        TestConfig { paging_type: PagingType::Paging5Level, address: 0x1000, size: FRAME_SIZE_4KB * 512 * 512 * 10 },
+    ];
 
     for test_config in test_configs {
         let TestConfig { size, address, paging_type } = test_config;
 
-        let num_pages = num_page_tables_required(address, size, paging_type).unwrap() + 0x3;
+        let num_pages = num_page_tables_required(address, size, paging_type).unwrap() + 0x4;
 
         let page_allocator = TestPageAllocator::new(num_pages, paging_type);
-
-        let pt = AArch64PageTable::new(page_allocator.clone(), paging_type);
+        let pt = X64PageTable::new(page_allocator.clone(), paging_type);
 
         assert!(pt.is_ok());
         let mut pt = pt.unwrap();
 
-        let attributes = MemoryAttributes::ReadOnly | MemoryAttributes::WriteCombining;
+        let attributes = MemoryAttributes::ReadOnly;
         let res = pt.map_memory_region(address, size, attributes);
         assert!(res.is_ok());
         assert_eq!(page_allocator.pages_allocated(), num_pages);
@@ -427,22 +567,24 @@ fn test_unmap_memory_address_0_to_ffff_ffff() {
         size: u64,
     }
 
-    let test_configs = [TestConfig { paging_type: PagingType::AArch64PageTable4KB, address: 0, size: FRAME_SIZE_4KB }];
+    let test_configs = [
+        TestConfig { paging_type: PagingType::Paging4Level, address: 0, size: FRAME_SIZE_4KB },
+        TestConfig { paging_type: PagingType::Paging5Level, address: 0, size: FRAME_SIZE_4KB },
+    ];
 
     for test_config in test_configs {
         let TestConfig { mut size, address, paging_type } = test_config;
 
         while size < 0xffff_ffff {
-            let num_pages = num_page_tables_required(address, size, paging_type).unwrap() + 0x3;
+            let num_pages = num_page_tables_required(address, size, paging_type).unwrap() + 0x4;
 
             let page_allocator = TestPageAllocator::new(num_pages, paging_type);
-
-            let pt = AArch64PageTable::new(page_allocator.clone(), paging_type);
+            let pt = X64PageTable::new(page_allocator.clone(), paging_type);
 
             assert!(pt.is_ok());
             let mut pt = pt.unwrap();
 
-            let attributes = MemoryAttributes::ReadOnly | MemoryAttributes::WriteCombining;
+            let attributes = MemoryAttributes::ReadOnly;
             let res = pt.map_memory_region(address, size, attributes);
             assert!(res.is_ok());
             assert_eq!(page_allocator.pages_allocated(), num_pages);
@@ -464,27 +606,34 @@ fn test_unmap_memory_address_single_page_from_0_to_ffff_ffff() {
         address_increment: u64,
     }
 
-    let test_configs = [TestConfig {
-        paging_type: PagingType::AArch64PageTable4KB,
-        address: 0,
-        address_increment: FRAME_SIZE_4KB,
-        size: FRAME_SIZE_4KB,
-    }];
+    let test_configs = [
+        TestConfig {
+            paging_type: PagingType::Paging4Level,
+            address: 0,
+            address_increment: FRAME_SIZE_4KB,
+            size: FRAME_SIZE_4KB,
+        },
+        TestConfig {
+            paging_type: PagingType::Paging5Level,
+            address: 0,
+            address_increment: FRAME_SIZE_4KB,
+            size: FRAME_SIZE_4KB,
+        },
+    ];
 
     for test_config in test_configs {
         let TestConfig { size, mut address, paging_type, address_increment } = test_config;
 
         while address < 0xffff_ffff {
-            let num_pages = num_page_tables_required(address, size, paging_type).unwrap() + 0x3;
+            let num_pages = num_page_tables_required(address, size, paging_type).unwrap() + 0x4;
 
             let page_allocator = TestPageAllocator::new(num_pages, paging_type);
-
-            let pt = AArch64PageTable::new(page_allocator.clone(), paging_type);
+            let pt = X64PageTable::new(page_allocator.clone(), paging_type);
 
             assert!(pt.is_ok());
             let mut pt = pt.unwrap();
 
-            let attributes = MemoryAttributes::ReadOnly | MemoryAttributes::WriteCombining;
+            let attributes = MemoryAttributes::ReadOnly;
             let res = pt.map_memory_region(address, size, attributes);
             assert!(res.is_ok());
 
@@ -504,27 +653,34 @@ fn test_unmap_memory_address_multiple_page_from_0_to_ffff_ffff() {
         address_increment: u64,
     }
 
-    let test_configs = [TestConfig {
-        paging_type: PagingType::AArch64PageTable4KB,
-        address: 0,
-        address_increment: FRAME_SIZE_4KB,
-        size: FRAME_SIZE_4KB << 1,
-    }];
+    let test_configs = [
+        TestConfig {
+            paging_type: PagingType::Paging4Level,
+            address: 0,
+            address_increment: FRAME_SIZE_4KB,
+            size: FRAME_SIZE_4KB << 1,
+        },
+        TestConfig {
+            paging_type: PagingType::Paging5Level,
+            address: 0,
+            address_increment: FRAME_SIZE_4KB,
+            size: FRAME_SIZE_4KB << 1,
+        },
+    ];
 
     for test_config in test_configs {
         let TestConfig { size, mut address, paging_type, address_increment } = test_config;
 
         while address < 0xffff_ffff {
-            let num_pages = num_page_tables_required(address, size, paging_type).unwrap() + 0x3;
+            let num_pages = num_page_tables_required(address, size, paging_type).unwrap() + 0x4;
 
             let page_allocator = TestPageAllocator::new(num_pages, paging_type);
-
-            let pt = AArch64PageTable::new(page_allocator.clone(), paging_type);
+            let pt = X64PageTable::new(page_allocator.clone(), paging_type);
 
             assert!(pt.is_ok());
             let mut pt = pt.unwrap();
 
-            let attributes = MemoryAttributes::ReadOnly | MemoryAttributes::WriteCombining;
+            let attributes = MemoryAttributes::ReadOnly;
             let res = pt.map_memory_region(address, size, attributes);
             assert!(res.is_ok());
             assert_eq!(page_allocator.pages_allocated(), num_pages);
@@ -544,7 +700,10 @@ fn test_unmap_memory_address_unaligned() {
         size: u64,
     }
 
-    let test_configs = [TestConfig { paging_type: PagingType::AArch64PageTable4KB, address: 0x1, size: 200 }];
+    let test_configs = [
+        TestConfig { paging_type: PagingType::Paging4Level, address: 0x1, size: 200 },
+        TestConfig { paging_type: PagingType::Paging5Level, address: 0x1, size: 200 },
+    ];
 
     for test_config in test_configs {
         let TestConfig { size, address, paging_type } = test_config;
@@ -552,8 +711,7 @@ fn test_unmap_memory_address_unaligned() {
         let max_pages: u64 = 10;
 
         let page_allocator = TestPageAllocator::new(max_pages, paging_type);
-
-        let pt = AArch64PageTable::new(page_allocator.clone(), paging_type);
+        let pt = X64PageTable::new(page_allocator.clone(), paging_type);
 
         assert!(pt.is_ok());
         let mut pt = pt.unwrap();
@@ -572,15 +730,17 @@ fn test_unmap_memory_address_zero_size() {
         size: u64,
     }
 
-    let test_configs = [TestConfig { paging_type: PagingType::AArch64PageTable4KB, address: 0x1000, size: 0 }];
+    let test_configs = [
+        TestConfig { paging_type: PagingType::Paging4Level, address: 0x1000, size: 0 },
+        TestConfig { paging_type: PagingType::Paging5Level, address: 0x1000, size: 0 },
+    ];
 
     for test_config in test_configs {
         let TestConfig { size, address, paging_type } = test_config;
         let max_pages: u64 = 10;
 
         let page_allocator = TestPageAllocator::new(max_pages, paging_type);
-
-        let pt = AArch64PageTable::new(page_allocator.clone(), paging_type);
+        let pt = X64PageTable::new(page_allocator.clone(), paging_type);
 
         assert!(pt.is_ok());
         let mut pt = pt.unwrap();
@@ -600,31 +760,30 @@ fn test_query_memory_address_simple() {
         size: u64,
     }
 
-    let test_configs = [TestConfig {
-        paging_type: PagingType::AArch64PageTable4KB,
-        address: 0x1000,
-        size: FRAME_SIZE_4KB * 512 * 512 * 10,
-    }];
+    let test_configs = [
+        TestConfig { paging_type: PagingType::Paging4Level, address: 0x1000, size: FRAME_SIZE_4KB * 512 * 512 * 10 },
+        TestConfig { paging_type: PagingType::Paging5Level, address: 0x1000, size: FRAME_SIZE_4KB * 512 * 512 * 10 },
+    ];
 
     for test_config in test_configs {
         let TestConfig { size, address, paging_type } = test_config;
 
-        let num_pages = num_page_tables_required(address, size, paging_type).unwrap() + 0x3;
+        let num_pages = num_page_tables_required(address, size, paging_type).unwrap() + 0x4;
 
         let page_allocator = TestPageAllocator::new(num_pages, paging_type);
-
-        let pt = AArch64PageTable::new(page_allocator.clone(), paging_type);
+        let pt = X64PageTable::new(page_allocator.clone(), paging_type);
 
         assert!(pt.is_ok());
         let mut pt = pt.unwrap();
 
-        let attributes = MemoryAttributes::ReadOnly | MemoryAttributes::WriteCombining;
+        let attributes = MemoryAttributes::ReadOnly;
         let res = pt.map_memory_region(address, size, attributes);
         assert!(res.is_ok());
         assert_eq!(page_allocator.pages_allocated(), num_pages);
 
         let res = pt.query_memory_region(address, size);
         assert!(res.is_ok());
+        assert_eq!(res.unwrap(), attributes);
     }
 }
 
@@ -636,28 +795,30 @@ fn test_query_memory_address_0_to_ffff_ffff() {
         size: u64,
     }
 
-    let test_configs =
-        [TestConfig { paging_type: PagingType::AArch64PageTable4KB, address: 0x1000, size: FRAME_SIZE_4KB }];
+    let test_configs = [
+        TestConfig { paging_type: PagingType::Paging4Level, address: 0x1000, size: FRAME_SIZE_4KB },
+        TestConfig { paging_type: PagingType::Paging5Level, address: 0x1000, size: FRAME_SIZE_4KB },
+    ];
 
     for test_config in test_configs {
         let TestConfig { mut size, address, paging_type } = test_config;
         while size < 0xffff_ffff {
-            let num_pages = num_page_tables_required(address, size, paging_type).unwrap() + 0x3;
+            let num_pages = num_page_tables_required(address, size, paging_type).unwrap() + 0x4;
 
             let page_allocator = TestPageAllocator::new(num_pages, paging_type);
-
-            let pt = AArch64PageTable::new(page_allocator.clone(), paging_type);
+            let pt = X64PageTable::new(page_allocator.clone(), paging_type);
 
             assert!(pt.is_ok());
             let mut pt = pt.unwrap();
 
-            let attributes = MemoryAttributes::ReadOnly | MemoryAttributes::WriteCombining;
+            let attributes = MemoryAttributes::ReadOnly;
             let res = pt.map_memory_region(address, size, attributes);
             assert!(res.is_ok());
             assert_eq!(page_allocator.pages_allocated(), num_pages);
 
             let res = pt.query_memory_region(address, size);
             assert!(res.is_ok());
+            assert_eq!(res.unwrap(), attributes);
             size <<= 1;
         }
     }
@@ -672,31 +833,29 @@ fn test_query_memory_address_single_page_from_0_to_ffff_ffff() {
         step: u64,
     }
 
-    let test_configs = [TestConfig {
-        paging_type: PagingType::AArch64PageTable4KB,
-        address: 0,
-        size: FRAME_SIZE_4KB,
-        step: FRAME_SIZE_4KB,
-    }];
+    let test_configs = [
+        TestConfig { paging_type: PagingType::Paging4Level, address: 0, size: FRAME_SIZE_4KB, step: FRAME_SIZE_4KB },
+        TestConfig { paging_type: PagingType::Paging5Level, address: 0, size: FRAME_SIZE_4KB, step: FRAME_SIZE_4KB },
+    ];
 
     for test_config in test_configs {
         let TestConfig { size, mut address, paging_type, step } = test_config;
         while address < 0xffff_ffff {
-            let num_pages = num_page_tables_required(address, size, paging_type).unwrap() + 0x3;
+            let num_pages = num_page_tables_required(address, size, paging_type).unwrap() + 0x4;
 
             let page_allocator = TestPageAllocator::new(num_pages, paging_type);
-
-            let pt = AArch64PageTable::new(page_allocator.clone(), paging_type);
+            let pt = X64PageTable::new(page_allocator.clone(), paging_type);
 
             assert!(pt.is_ok());
             let mut pt = pt.unwrap();
 
-            let attributes = MemoryAttributes::ReadOnly | MemoryAttributes::WriteCombining;
+            let attributes = MemoryAttributes::ReadOnly;
             let res = pt.map_memory_region(address, size, attributes);
             assert!(res.is_ok());
 
             let res = pt.query_memory_region(address, size);
             assert!(res.is_ok());
+            assert_eq!(res.unwrap(), attributes);
             address += step;
         }
     }
@@ -711,32 +870,40 @@ fn test_query_memory_address_multiple_page_from_0_to_ffff_ffff() {
         step: u64,
     }
 
-    let test_configs = [TestConfig {
-        paging_type: PagingType::AArch64PageTable4KB,
-        address: 0,
-        size: FRAME_SIZE_4KB << 1,
-        step: FRAME_SIZE_4KB,
-    }];
+    let test_configs = [
+        TestConfig {
+            paging_type: PagingType::Paging4Level,
+            address: 0,
+            size: FRAME_SIZE_4KB << 1,
+            step: FRAME_SIZE_4KB,
+        },
+        TestConfig {
+            paging_type: PagingType::Paging5Level,
+            address: 0,
+            size: FRAME_SIZE_4KB << 1,
+            step: FRAME_SIZE_4KB,
+        },
+    ];
 
     for test_config in test_configs {
         let TestConfig { size, mut address, paging_type, step } = test_config;
         while address < 0xffff_ffff {
-            let num_pages = num_page_tables_required(address, size, paging_type).unwrap() + 0x3;
+            let num_pages = num_page_tables_required(address, size, paging_type).unwrap() + 0x4;
 
             let page_allocator = TestPageAllocator::new(num_pages, paging_type);
-
-            let pt = AArch64PageTable::new(page_allocator.clone(), paging_type);
+            let pt = X64PageTable::new(page_allocator.clone(), paging_type);
 
             assert!(pt.is_ok());
             let mut pt = pt.unwrap();
 
-            let attributes = MemoryAttributes::ReadOnly | MemoryAttributes::WriteCombining;
+            let attributes = MemoryAttributes::ReadOnly;
             let res = pt.map_memory_region(address, size, attributes);
             assert!(res.is_ok());
             assert_eq!(page_allocator.pages_allocated(), num_pages);
 
             let res = pt.query_memory_region(address, size);
             assert!(res.is_ok());
+            assert_eq!(res.unwrap(), attributes);
             address += step;
         }
     }
@@ -746,9 +913,9 @@ fn test_query_memory_address_multiple_page_from_0_to_ffff_ffff() {
 fn test_query_memory_address_unaligned() {
     let max_pages: u64 = 10;
 
-    let page_allocator = TestPageAllocator::new(max_pages, PagingType::AArch64PageTable4KB);
+    let page_allocator = TestPageAllocator::new(max_pages, PagingType::Paging4Level);
 
-    let pt = AArch64PageTable::new(page_allocator.clone(), PagingType::AArch64PageTable4KB);
+    let pt = X64PageTable::new(page_allocator.clone(), PagingType::Paging4Level);
 
     assert!(pt.is_ok());
     let pt = pt.unwrap();
@@ -764,9 +931,9 @@ fn test_query_memory_address_unaligned() {
 fn test_query_memory_address_zero_size() {
     let max_pages: u64 = 10;
 
-    let page_allocator = TestPageAllocator::new(max_pages, PagingType::AArch64PageTable4KB);
+    let page_allocator = TestPageAllocator::new(max_pages, PagingType::Paging4Level);
 
-    let pt = AArch64PageTable::new(page_allocator.clone(), PagingType::AArch64PageTable4KB);
+    let pt = X64PageTable::new(page_allocator.clone(), PagingType::Paging4Level);
 
     assert!(pt.is_ok());
     let pt = pt.unwrap();
@@ -786,7 +953,10 @@ fn test_query_memory_address_inconsistent_mappings() {
         size: u64,
     }
 
-    let test_configs = [TestConfig { paging_type: PagingType::AArch64PageTable4KB, address: 0x1000, size: 0x3000 }];
+    let test_configs = [
+        TestConfig { paging_type: PagingType::Paging4Level, address: 0x1000, size: 0x3000 },
+        TestConfig { paging_type: PagingType::Paging5Level, address: 0x1000, size: 0x3000 },
+    ];
 
     for test_config in test_configs {
         let TestConfig { size, address, paging_type } = test_config;
@@ -794,13 +964,13 @@ fn test_query_memory_address_inconsistent_mappings() {
         let num_pages = num_page_tables_required(address, size, paging_type).unwrap() + 0x4;
 
         let page_allocator = TestPageAllocator::new(num_pages, paging_type);
-        let pt = AArch64PageTable::new(page_allocator.clone(), paging_type);
+        let pt = X64PageTable::new(page_allocator.clone(), paging_type);
 
         assert!(pt.is_ok());
         let mut pt = pt.unwrap();
 
         // Map the first part of the range
-        let attributes = MemoryAttributes::ReadOnly | MemoryAttributes::Writeback;
+        let attributes = MemoryAttributes::ReadOnly;
         let res = pt.map_memory_region(address, FRAME_SIZE_4KB, attributes);
         assert!(res.is_ok());
 
@@ -823,19 +993,22 @@ fn test_query_memory_address_inconsistent_mappings_across_2mb_boundary() {
         size: u64,
     }
 
-    let test_configs = [TestConfig { paging_type: PagingType::AArch64PageTable4KB, address: 0x0, size: 0x400000 }];
+    let test_configs = [
+        TestConfig { paging_type: PagingType::Paging4Level, address: 0x0, size: 0x400000 },
+        TestConfig { paging_type: PagingType::Paging5Level, address: 0x0, size: 0x400000 },
+    ];
 
     for test_config in test_configs {
         let TestConfig { size, address, paging_type } = test_config;
 
         let page_allocator = TestPageAllocator::new(0x1000, paging_type);
-        let pt = AArch64PageTable::new(page_allocator.clone(), paging_type);
+        let pt = X64PageTable::new(page_allocator.clone(), paging_type);
 
         assert!(pt.is_ok());
         let mut pt = pt.unwrap();
 
         // Map the first 2MB, but not the second 2MB, map in 1MB chunks so that all PTEs are mapped
-        let attributes = MemoryAttributes::ReadOnly | MemoryAttributes::Writeback;
+        let attributes = MemoryAttributes::ReadOnly;
         let res = pt.map_memory_region(address, 0x100000, attributes);
         assert!(res.is_ok());
         let res = pt.map_memory_region(address + 0x100000, 0x100000, attributes);
@@ -880,30 +1053,28 @@ fn test_remap_memory_address_simple() {
         size: u64,
     }
 
-    let test_configs = [TestConfig {
-        paging_type: PagingType::AArch64PageTable4KB,
-        address: 0x1000,
-        size: FRAME_SIZE_4KB * 512 * 512 * 10,
-    }];
+    let test_configs = [
+        TestConfig { paging_type: PagingType::Paging4Level, address: 0x1000, size: FRAME_SIZE_4KB * 512 * 512 * 10 },
+        TestConfig { paging_type: PagingType::Paging5Level, address: 0x1000, size: FRAME_SIZE_4KB * 512 * 512 * 10 },
+    ];
 
     for test_config in test_configs {
         let TestConfig { size, address, paging_type } = test_config;
 
-        let num_pages = num_page_tables_required(address, size, paging_type).unwrap() + 0x3;
+        let num_pages = num_page_tables_required(address, size, paging_type).unwrap() + 0x4;
 
         let page_allocator = TestPageAllocator::new(num_pages, paging_type);
-
-        let pt = AArch64PageTable::new(page_allocator.clone(), paging_type);
+        let pt = X64PageTable::new(page_allocator.clone(), paging_type);
 
         assert!(pt.is_ok());
         let mut pt = pt.unwrap();
 
-        let attributes = MemoryAttributes::ReadOnly | MemoryAttributes::WriteCombining;
+        let attributes = MemoryAttributes::ReadOnly;
         let res = pt.map_memory_region(address, size, attributes);
         assert!(res.is_ok());
         assert_eq!(page_allocator.pages_allocated(), num_pages);
 
-        let attributes = MemoryAttributes::ExecuteProtect | MemoryAttributes::Writeback;
+        let attributes = MemoryAttributes::ExecuteProtect;
         let res = pt.remap_memory_region(address, size, attributes);
         assert!(res.is_ok());
     }
@@ -917,27 +1088,29 @@ fn test_remap_memory_address_0_to_ffff_ffff() {
         size: u64,
     }
 
-    let test_configs = [TestConfig { paging_type: PagingType::AArch64PageTable4KB, address: 0, size: FRAME_SIZE_4KB }];
+    let test_configs = [
+        TestConfig { paging_type: PagingType::Paging4Level, address: 0, size: FRAME_SIZE_4KB },
+        TestConfig { paging_type: PagingType::Paging5Level, address: 0, size: FRAME_SIZE_4KB },
+    ];
 
     for test_config in test_configs {
         let TestConfig { mut size, address, paging_type } = test_config;
 
         while size < 0xffff_ffff {
-            let num_pages = num_page_tables_required(address, size, paging_type).unwrap() + 0x3;
+            let num_pages = num_page_tables_required(address, size, paging_type).unwrap() + 0x4;
 
             let page_allocator = TestPageAllocator::new(num_pages, paging_type);
-
-            let pt = AArch64PageTable::new(page_allocator.clone(), paging_type);
+            let pt = X64PageTable::new(page_allocator.clone(), paging_type);
 
             assert!(pt.is_ok());
             let mut pt = pt.unwrap();
 
-            let attributes = MemoryAttributes::ReadOnly | MemoryAttributes::WriteCombining;
+            let attributes = MemoryAttributes::ReadOnly;
             let res = pt.map_memory_region(address, size, attributes);
             assert!(res.is_ok());
             assert_eq!(page_allocator.pages_allocated(), num_pages);
 
-            let attributes = MemoryAttributes::ExecuteProtect | MemoryAttributes::Writeback;
+            let attributes = MemoryAttributes::ExecuteProtect;
             let res = pt.remap_memory_region(address, size, attributes);
             assert!(res.is_ok());
             size <<= 1;
@@ -955,31 +1128,38 @@ fn test_remap_memory_address_single_page_from_0_to_ffff_ffff() {
         address_increment: u64,
     }
 
-    let test_configs = [TestConfig {
-        paging_type: PagingType::AArch64PageTable4KB,
-        address: 0,
-        address_increment: FRAME_SIZE_4KB,
-        size: FRAME_SIZE_4KB,
-    }];
+    let test_configs = [
+        TestConfig {
+            paging_type: PagingType::Paging4Level,
+            address: 0,
+            address_increment: FRAME_SIZE_4KB,
+            size: FRAME_SIZE_4KB,
+        },
+        TestConfig {
+            paging_type: PagingType::Paging5Level,
+            address: 0,
+            address_increment: FRAME_SIZE_4KB,
+            size: FRAME_SIZE_4KB,
+        },
+    ];
 
     for test_config in test_configs {
         let TestConfig { size, mut address, address_increment, paging_type } = test_config;
 
         while address < 0xffff_ffff {
-            let num_pages = num_page_tables_required(address, size, paging_type).unwrap() + 0x3;
+            let num_pages = num_page_tables_required(address, size, paging_type).unwrap() + 0x4;
 
             let page_allocator = TestPageAllocator::new(num_pages, paging_type);
-
-            let pt = AArch64PageTable::new(page_allocator.clone(), paging_type);
+            let pt = X64PageTable::new(page_allocator.clone(), paging_type);
 
             assert!(pt.is_ok());
             let mut pt = pt.unwrap();
 
-            let attributes = MemoryAttributes::ReadOnly | MemoryAttributes::WriteCombining;
+            let attributes = MemoryAttributes::ReadOnly;
             let res = pt.map_memory_region(address, size, attributes);
             assert!(res.is_ok());
 
-            let attributes = MemoryAttributes::ExecuteProtect | MemoryAttributes::Writeback;
+            let attributes = MemoryAttributes::ExecuteProtect;
             let res = pt.remap_memory_region(address, size, attributes);
             assert!(res.is_ok());
             address += address_increment;
@@ -996,32 +1176,39 @@ fn test_remap_memory_address_multiple_page_from_0_to_ffff_ffff() {
         address_increment: u64,
     }
 
-    let test_configs = [TestConfig {
-        paging_type: PagingType::AArch64PageTable4KB,
-        address: 0,
-        address_increment: FRAME_SIZE_4KB,
-        size: FRAME_SIZE_4KB << 1,
-    }];
+    let test_configs = [
+        TestConfig {
+            paging_type: PagingType::Paging4Level,
+            address: 0,
+            address_increment: FRAME_SIZE_4KB,
+            size: FRAME_SIZE_4KB << 1,
+        },
+        TestConfig {
+            paging_type: PagingType::Paging5Level,
+            address: 0,
+            address_increment: FRAME_SIZE_4KB,
+            size: FRAME_SIZE_4KB << 1,
+        },
+    ];
 
     for test_config in test_configs {
         let TestConfig { size, mut address, address_increment, paging_type } = test_config;
 
         while address < 0xffff_ffff {
-            let num_pages = num_page_tables_required(address, size, paging_type).unwrap() + 0x3;
+            let num_pages = num_page_tables_required(address, size, paging_type).unwrap() + 0x4;
 
             let page_allocator = TestPageAllocator::new(num_pages, paging_type);
-
-            let pt = AArch64PageTable::new(page_allocator.clone(), paging_type);
+            let pt = X64PageTable::new(page_allocator.clone(), paging_type);
 
             assert!(pt.is_ok());
             let mut pt = pt.unwrap();
 
-            let attributes = MemoryAttributes::ReadOnly | MemoryAttributes::WriteCombining;
+            let attributes = MemoryAttributes::ReadOnly;
             let res = pt.map_memory_region(address, size, attributes);
             assert!(res.is_ok());
             assert_eq!(page_allocator.pages_allocated(), num_pages);
 
-            let attributes = MemoryAttributes::ExecuteProtect | MemoryAttributes::Writeback;
+            let attributes = MemoryAttributes::ExecuteProtect;
             let res = pt.remap_memory_region(address, size, attributes);
             assert!(res.is_ok());
             address += address_increment;
@@ -1037,7 +1224,10 @@ fn test_remap_memory_address_unaligned() {
         size: u64,
     }
 
-    let test_configs = [TestConfig { paging_type: PagingType::AArch64PageTable4KB, address: 0x1, size: 200 }];
+    let test_configs = [
+        TestConfig { paging_type: PagingType::Paging4Level, address: 0x1, size: 200 },
+        TestConfig { paging_type: PagingType::Paging5Level, address: 0x1, size: 200 },
+    ];
 
     for test_config in test_configs {
         let TestConfig { size, address, paging_type } = test_config;
@@ -1045,12 +1235,12 @@ fn test_remap_memory_address_unaligned() {
 
         let page_allocator = TestPageAllocator::new(max_pages, paging_type);
 
-        let pt = AArch64PageTable::new(page_allocator.clone(), paging_type);
+        let pt = X64PageTable::new(page_allocator.clone(), paging_type);
 
         assert!(pt.is_ok());
         let mut pt = pt.unwrap();
 
-        let attributes = MemoryAttributes::ExecuteProtect | MemoryAttributes::Writeback;
+        let attributes = MemoryAttributes::ExecuteProtect;
         let res = pt.remap_memory_region(address, size, attributes);
         assert!(res.is_err());
         assert_eq!(res, Err(PtError::UnalignedAddress));
@@ -1065,7 +1255,10 @@ fn test_remap_memory_address_zero_size() {
         size: u64,
     }
 
-    let test_configs = [TestConfig { paging_type: PagingType::AArch64PageTable4KB, address: 0x1000, size: 0 }];
+    let test_configs = [
+        TestConfig { paging_type: PagingType::Paging4Level, address: 0x1000, size: 0 },
+        TestConfig { paging_type: PagingType::Paging5Level, address: 0x1000, size: 0 },
+    ];
 
     for test_config in test_configs {
         let TestConfig { size, address, paging_type } = test_config;
@@ -1074,12 +1267,12 @@ fn test_remap_memory_address_zero_size() {
 
         let page_allocator = TestPageAllocator::new(max_pages, paging_type);
 
-        let pt = AArch64PageTable::new(page_allocator.clone(), paging_type);
+        let pt = X64PageTable::new(page_allocator.clone(), paging_type);
 
         assert!(pt.is_ok());
         let mut pt = pt.unwrap();
 
-        let attributes = MemoryAttributes::ExecuteProtect | MemoryAttributes::Writeback;
+        let attributes = MemoryAttributes::ExecuteProtect;
         let res = pt.remap_memory_region(address, size, attributes);
         assert!(res.is_err());
         assert_eq!(res, Err(PtError::InvalidMemoryRange));
@@ -1094,32 +1287,30 @@ fn test_from_existing_page_table() {
         size: u64,
     }
 
-    let test_configs = [TestConfig {
-        paging_type: PagingType::AArch64PageTable4KB,
-        address: 0x1000,
-        size: FRAME_SIZE_4KB * 512 * 512 * 10,
-    }];
+    let test_configs = [
+        TestConfig { paging_type: PagingType::Paging4Level, address: 0x1000, size: FRAME_SIZE_4KB * 512 * 512 * 10 },
+        TestConfig { paging_type: PagingType::Paging5Level, address: 0x1000, size: FRAME_SIZE_4KB * 512 * 512 * 10 },
+    ];
 
     for test_config in test_configs {
         let TestConfig { size, address, paging_type } = test_config;
 
-        let num_pages = num_page_tables_required(address, size, paging_type).unwrap() + 0x3;
+        let num_pages = num_page_tables_required(address, size, paging_type).unwrap() + 0x4;
 
         let page_allocator = TestPageAllocator::new(num_pages, paging_type);
-
-        let pt = AArch64PageTable::new(page_allocator.clone(), paging_type);
+        let pt = X64PageTable::new(page_allocator.clone(), paging_type);
 
         assert!(pt.is_ok());
         let mut pt = pt.unwrap();
 
-        let attributes = MemoryAttributes::ReadOnly | MemoryAttributes::WriteCombining;
+        let attributes = MemoryAttributes::ReadOnly;
         let res = pt.map_memory_region(address, size, attributes);
         assert!(res.is_ok());
         assert_eq!(page_allocator.pages_allocated(), num_pages);
 
         // Create a new page table from the existing one
         let new_pt =
-            unsafe { AArch64PageTable::from_existing(pt.into_page_table_root(), page_allocator.clone(), paging_type) };
+            unsafe { X64PageTable::from_existing(pt.into_page_table_root(), page_allocator.clone(), paging_type) };
         assert!(new_pt.is_ok());
         let new_pt = new_pt.unwrap();
 
@@ -1137,20 +1328,22 @@ fn test_dump_page_tables() {
         size: u64,
     }
 
-    let test_configs = [TestConfig { paging_type: PagingType::AArch64PageTable4KB, address: 0, size: 0x4000 }];
+    let test_configs = [TestConfig { paging_type: PagingType::Paging4Level, address: 0, size: 0x8000 }];
+
+    set_logger();
 
     for test_config in test_configs {
         let TestConfig { size, address, paging_type } = test_config;
 
-        let num_pages = num_page_tables_required(address, size, paging_type).unwrap() + 0x3;
+        let num_pages = num_page_tables_required(address, size, paging_type).unwrap() + 0x4;
 
         let page_allocator = TestPageAllocator::new(num_pages, paging_type);
-        let pt = AArch64PageTable::new(page_allocator.clone(), paging_type);
+        let pt = X64PageTable::new(page_allocator.clone(), paging_type);
 
         assert!(pt.is_ok());
         let mut pt = pt.unwrap();
 
-        let attributes = MemoryAttributes::ReadOnly | MemoryAttributes::WriteCombining;
+        let attributes = MemoryAttributes::ReadOnly;
         let res = pt.map_memory_region(address, size, attributes);
         assert!(res.is_ok());
 
@@ -1180,31 +1373,37 @@ fn test_large_page_splitting() {
 
     let test_configs = [
         TestConfig {
-            paging_type: PagingType::AArch64PageTable4KB,
+            paging_type: PagingType::Paging4Level,
             mapped_range: TestRange { address: 0, size: 0x200000 },
             split_range: TestRange { address: 0, size: 0x1000 },
             page_increase: 1, // 1 2MB page split into 4KB pages, adds 1 PT.
         },
         TestConfig {
-            paging_type: PagingType::AArch64PageTable4KB,
+            paging_type: PagingType::Paging5Level,
+            mapped_range: TestRange { address: 0, size: 0x200000 },
+            split_range: TestRange { address: 0, size: 0x1000 },
+            page_increase: 1, // 1 2MB page split into 4KB pages, adds 1 PT.
+        },
+        TestConfig {
+            paging_type: PagingType::Paging4Level,
             mapped_range: TestRange { address: 0, size: 0x600000 },
             split_range: TestRange { address: 0x100000, size: 0x400000 },
             page_increase: 2, // 3 2MB pages split into 1 2MB with 4KB on either side, adds 2 PT.
         },
         TestConfig {
-            paging_type: PagingType::AArch64PageTable4KB,
+            paging_type: PagingType::Paging4Level,
             mapped_range: TestRange { address: 0, size: 0x40000000 },
             split_range: TestRange { address: 0, size: 0x1000 },
             page_increase: 2, // 1 1GB page split into 4KB + 2MB pages, adds 1PD + 1 PT.
         },
         TestConfig {
-            paging_type: PagingType::AArch64PageTable4KB,
+            paging_type: PagingType::Paging4Level,
             mapped_range: TestRange { address: 0, size: 0x40000000 },
             split_range: TestRange { address: 0, size: 0x200000 },
             page_increase: 1, // 1 1GB page split into 2MB pages, adds 1PD.
         },
         TestConfig {
-            paging_type: PagingType::AArch64PageTable4KB,
+            paging_type: PagingType::Paging4Level,
             mapped_range: TestRange { address: 0, size: 0x40000000 },
             split_range: TestRange { address: 0x1FF000, size: 0x2000 },
             page_increase: 3, // 1 1GB page split into 4KB + 2MB pages along 2 2MB pages, adds 1PD + 2 PT.
@@ -1216,18 +1415,18 @@ fn test_large_page_splitting() {
         Remap,
     }
 
-    let orig_attributes = MemoryAttributes::empty() | MemoryAttributes::Writeback;
-    let remap_attributes = MemoryAttributes::ExecuteProtect | MemoryAttributes::Writeback;
+    let orig_attributes = MemoryAttributes::empty();
+    let remap_attributes = MemoryAttributes::ExecuteProtect;
 
     // Test the splitting when remapping.
     for test_config in test_configs {
         let TestConfig { paging_type, mapped_range, split_range, page_increase } = test_config;
         for action in [TestAction::Unmap, TestAction::Remap] {
             let num_pages =
-                num_page_tables_required(mapped_range.address, mapped_range.size, paging_type).unwrap() + 0x3;
+                num_page_tables_required(mapped_range.address, mapped_range.size, paging_type).unwrap() + 0x4;
 
             let page_allocator = TestPageAllocator::new(num_pages + page_increase, paging_type);
-            let pt = AArch64PageTable::new(page_allocator.clone(), paging_type);
+            let pt = X64PageTable::new(page_allocator.clone(), paging_type);
 
             assert!(pt.is_ok());
             let mut pt = pt.unwrap();
@@ -1269,32 +1468,48 @@ fn test_large_page_splitting() {
 fn test_self_map() {
     struct TestConfig {
         paging_type: PagingType,
+        address: u64,
     }
 
-    let test_configs = [TestConfig { paging_type: PagingType::AArch64PageTable4KB }];
+    let test_configs = [
+        TestConfig { paging_type: PagingType::Paging4Level, address: 0x1000 },
+        TestConfig { paging_type: PagingType::Paging5Level, address: 0x1000 },
+    ];
 
     for test_config in test_configs {
-        let TestConfig { paging_type } = test_config;
+        let TestConfig { address, paging_type } = test_config;
 
         let page_allocator = TestPageAllocator::new(0x1000, paging_type);
-        let pt = AArch64PageTable::new(page_allocator.clone(), paging_type);
+        let pt = X64PageTable::new(page_allocator.clone(), paging_type);
 
         assert!(pt.is_ok());
-        let pt = pt.unwrap();
+        let mut pt = pt.unwrap();
 
-        // query self map base addresses
-        let res = pt.query_memory_region(FOUR_LEVEL_PML4_SELF_MAP_BASE, FRAME_SIZE_4KB);
+        // Create some pages before the install, so VA = PA for accessing
+        for i in 0..10 {
+            let test_address = address + i * FRAME_SIZE_4KB;
+            let test_size = FRAME_SIZE_4KB;
+            let test_attributes = match i % 3 {
+                0 => MemoryAttributes::ReadOnly,
+                1 => MemoryAttributes::empty(),
+                _ => MemoryAttributes::ExecuteProtect,
+            };
+
+            let res = pt.map_memory_region(test_address, test_size, test_attributes);
+            assert!(res.is_ok());
+
+            let res = pt.query_memory_region(test_address, test_size);
+            assert!(res.is_ok());
+            assert_eq!(res.unwrap(), test_attributes);
+        }
+
+        let res = pt.install_page_table();
         assert!(res.is_ok());
-
-        // we can't query the zero VA because in new() it is not mapped on purpose, so we just check we mapped
-        // down to the PTE level
-        let res = pt.query_memory_region(ZERO_VA_4_LEVEL, FRAME_SIZE_4KB);
-        assert!(res.is_err());
 
         let root = pt.into_page_table_root();
 
         // now we should see the zero VA and the self map entries in the base page table
-        let zero_va_top_level = root + ZERO_VA_INDEX * size_of::<AArch64Descriptor>() as u64;
+        let zero_va_top_level = root + ZERO_VA_INDEX * size_of::<PageTableEntry>() as u64;
         assert!(unsafe { *(zero_va_top_level as *const u64) != 0 });
 
         let zero_va_pdp_level = unsafe { *(zero_va_top_level as *const u64) & !0xFFF };
@@ -1306,11 +1521,25 @@ fn test_self_map() {
         let zero_va_pt_level = unsafe { *(zero_va_pd_level as *const u64) & !0xFFF };
         assert!(unsafe { *(zero_va_pt_level as *const u64) != 0 });
 
-        let zero_va_pa = unsafe { *(zero_va_pt_level as *const u64) & !0xFFF };
-        assert!(zero_va_pa == 0);
+        match paging_type {
+            PagingType::Paging4Level => {
+                // 4 level paging ends here, we expect the zero VA to be unmapped
+                let zero_va_pa = unsafe { *(zero_va_pt_level as *const u64) & !0xFFF };
+                assert!(zero_va_pa == 0);
+            }
+            PagingType::Paging5Level => {
+                // 5 level paging has another level to it, so names above aren't correct
+                let zero_va_real_pt_level = unsafe { *(zero_va_pt_level as *const u64) & !0xFFF };
+                assert!(unsafe { *(zero_va_pt_level as *const u64) != 0 });
 
-        let self_map_top_level = root + SELF_MAP_INDEX * size_of::<AArch64Descriptor>() as u64;
-        assert_eq!((unsafe { *(self_map_top_level as *const u64) } & !0xFFF), root);
+                let zero_va_pa_level = unsafe { *(zero_va_real_pt_level as *const u64) & !0xFFF };
+                assert!(zero_va_pa_level == 0);
+            }
+            _ => panic!("Invalid paging type"),
+        }
+
+        let self_map_top_level = root + SELF_MAP_INDEX * size_of::<PageTableEntry>() as u64;
+        assert_eq!(unsafe { *(self_map_top_level as *const u64) } & CR3_PAGE_BASE_ADDRESS_MASK, root);
     }
 }
 
@@ -1321,13 +1550,16 @@ fn test_install_page_table() {
         address: u64,
     }
 
-    let test_configs = [TestConfig { paging_type: PagingType::AArch64PageTable4KB, address: 0x1000 }];
+    let test_configs = [
+        TestConfig { paging_type: PagingType::Paging4Level, address: 0x1000 },
+        TestConfig { paging_type: PagingType::Paging5Level, address: 0x1000 },
+    ];
 
     for test_config in test_configs {
         let TestConfig { address, paging_type } = test_config;
 
         let page_allocator = TestPageAllocator::new(0x1000, paging_type);
-        let pt = AArch64PageTable::new(page_allocator.clone(), paging_type);
+        let pt = X64PageTable::new(page_allocator.clone(), paging_type);
 
         assert!(pt.is_ok());
         let mut pt = pt.unwrap();
@@ -1337,9 +1569,9 @@ fn test_install_page_table() {
             let test_address = address + i * FRAME_SIZE_4KB;
             let test_size = FRAME_SIZE_4KB;
             let test_attributes = match i % 3 {
-                0 => MemoryAttributes::ReadOnly | MemoryAttributes::Writeback,
-                1 => MemoryAttributes::Writeback,
-                _ => MemoryAttributes::ExecuteProtect | MemoryAttributes::Writeback,
+                0 => MemoryAttributes::ReadOnly,
+                1 => MemoryAttributes::empty(),
+                _ => MemoryAttributes::ExecuteProtect,
             };
 
             let res = pt.map_memory_region(test_address, test_size, test_attributes);
@@ -1358,9 +1590,9 @@ fn test_install_page_table() {
             let test_address = address + i * FRAME_SIZE_4KB;
             let test_size = FRAME_SIZE_4KB;
             let test_attributes = match i % 3 {
-                0 => MemoryAttributes::ReadOnly | MemoryAttributes::Writeback,
-                1 => MemoryAttributes::Writeback,
-                _ => MemoryAttributes::ExecuteProtect | MemoryAttributes::Writeback,
+                0 => MemoryAttributes::ReadOnly,
+                1 => MemoryAttributes::empty(),
+                _ => MemoryAttributes::ExecuteProtect,
             };
 
             let res = pt.map_memory_region(test_address, test_size, test_attributes);
@@ -1383,9 +1615,9 @@ fn test_install_page_table() {
             let test_address = address + i * FRAME_SIZE_4KB;
             let test_size = FRAME_SIZE_4KB;
             let test_attributes = match i % 3 {
-                0 => MemoryAttributes::ReadOnly | MemoryAttributes::Writeback,
-                1 => MemoryAttributes::Writeback,
-                _ => MemoryAttributes::ExecuteProtect | MemoryAttributes::Writeback,
+                0 => MemoryAttributes::ReadOnly,
+                1 => MemoryAttributes::empty(),
+                _ => MemoryAttributes::ExecuteProtect,
             };
 
             let res = pt.remap_memory_region(test_address, test_size, test_attributes);
